+++ src/services/app.service.ts
@@ -3,6 +3,7 @@ import { Injectable } from '@nestjs/common';
 @Injectable()
 export class AppService {
   getHello(): string {
-    return 'Hello World!!!';
+    let word: string = 'hee'
+    return 'Hello World!' + word;
   }
 }
